refactor: extract `useInvalidateOnSuccess` pattern
Trying to help prevent a common mistake, which is mutating without
invalidating the correct API responses from the cache.

This approach lifts up mutation another level into a hook that uses
`useInvalidateOnSuccess`. `useInvalidateOnSuccess` accepts a list of
invalidation functions to call once your mutation succeeds and then
gives access to the mutator function and pending state to be used as
usual in our forms.